+++ header.php
@@ -6,7 +6,7 @@
  * @since   1.0
  * @author  RewindCreation
  * @license GPL v3 or later
- * @link    http://www.rewindcreation.com/
+ * @link    http://rewindcreation.com/
  */
 ?>
 <!DOCTYPE html>
@@ -15,9 +15,9 @@
 	<meta charset="<?php bloginfo( 'charset' ); ?>">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="profile" href="http://gmpg.org/xfn/11">
-	<?php if ( is_singular() && pings_open( get_queried_object() ) ) : ?>
+	<?php if ( is_singular() && pings_open( get_queried_object() ) ) { ?>
 	<link rel="pingback" href="<?php bloginfo( 'pingback_url' ); ?>">
-	<?php endif; ?>
+	<?php } ?>
 	<?php wp_head(); ?>
 </head>
 <body <?php body_class(); ?>>
@@ -25,7 +25,7 @@
 <div class="off-canvas-wrapper-inner" data-off-canvas-wrapper>
 	<?php get_template_part( 'parts/left', 'menu' ); ?>
 <div id="wrapper" class="site off-canvas-content" data-off-canvas-content>
-			<a class="skip-link screen-reader-text" href="#content"><?php _e( 'Skip to content', 'hana' ); ?></a>
+	<a class="skip-link screen-reader-text" href="#content"><?php _e( 'Skip to content', 'hana' ); ?></a>
 	<?php do_action('hana_header_top'); //Action Hook ?>
  	<header id="masthead" class="site-header clearfix" role="banner">
 		<?php get_template_part( 'parts/top', 'menu' ); ?>
@@ -33,28 +33,9 @@
  <?php
 	if ( 'full' == hana_option( 'slider_type' ) )
 		hana_featured_top();
-	if (  has_nav_menu( 'section' ) ||  ( hana_option( 'social_section' ) && has_nav_menu( 'social' ) ) ) {
-?>
-		<div class="sectionmenu show-for-large">
-			<div class="column row <?php if ( hana_option( 'fullwidth_header' ) ) echo 'expanded'; ?>">
-<?php
-			if (  has_nav_menu( 'section' ) ) { ?>
-				<nav class="section-menu">
-<?php				wp_nav_menu( array(
-						'theme_location'  => 'section',
-						'menu_class' => '',	
-						'container'  => false,
-					)); ?>
-				</nav>
-<?php		}
-		if ( hana_option( 'social_section') && has_nav_menu( 'social' ) )
-			hana_social_menu( 'social social-section float-right' ); ?>
-			</div>
-		</div>
-<?php
-	}
+	get_template_part( 'parts/section', 'menu' );
 	if ( 'full' != hana_option( 'slider_type' ) )
 		hana_featured_top();
 	do_action('hana_header_before_main'); //Action Hook
 ?>
-<div id="main" class="row<?php hana_full_width_class() ?>">
+<div id="main" class="<?php echo hana_main_class();?>">