+++ sidebar.php
@@ -9,13 +9,13 @@
  * @link    http://rewindcreation.com/
  */
 	$sidebar_pos = hana_option( 'sidebar_pos' );
-	$sidebar1 = intval( hana_option( 'sidebar1_column' ) );
-	$sidebar2 = intval( hana_option( 'sidebar2_column' ) );
+	$sidebar1 = absint( hana_option( 'sidebar1_column' ) );
+	$sidebar2 = absint( hana_option( 'sidebar2_column' ) );
 	$width = $sidebar1 + $sidebar2;
 	//Sticky Sidebar
 	$sticky = '';
 	if ( hana_option( 'sticky_sidebar' ) )
-		$sticky = '<div class="sticky" data-sticky data-anchor="content" data-margin-top="4" style="width:100%">';
+		$sticky = '<div class="sticky sticky-sidebar" data-sticky data-anchor="content" data-margin-top="4">';
 	// Full Sidebar	
 	if ( 'both' != $sidebar_pos && 'none' != $sidebar_pos && $width > 0 && is_active_sidebar( 'sidebar-full' ) ) { ?>
 		<aside id="sidebar-full" class="<?php echo hana_sidebar_class( 'full' ); ?> sidebar" role="complementary" data-sticky-container>