+++ changelog.md
@@ -6,7 +6,9 @@
 * Added option not to shrink the topbar
 * Improved widget CSS classes
 * Added option to align slider to top
-* Changed social link to use menu
+* Changed social link to use Social menu
+* Added Copyright text option
+* Added Rewind Scheme
  
 ### 1.0.3